Address review comments from nc6
