+++ cardano-crypto-class/src/Cardano/Crypto/KES/Sum.hs
@@ -49,6 +49,7 @@ import           NoThunks.Class (NoThunks)
 
 import           Cardano.Binary (FromCBOR (..), ToCBOR (..))
 
+import           Cardano.Crypto.Util
 import           Cardano.Crypto.Seed
 import           Cardano.Crypto.Hash.Class
 import           Cardano.Crypto.KES.Class
@@ -277,31 +278,6 @@ instance (KESAlgorithm d, HashAlgorithm h, Typeable d)
         off_vk0    = size_sig
         off_vk1    = off_vk0 + size_vk
 
-hashPairOfVKeys :: (KESAlgorithm d, HashAlgorithm h)
-                => (VerKeyKES d, VerKeyKES d)
-                -> Hash h (VerKeyKES d, VerKeyKES d)
-hashPairOfVKeys =
-    hashWith $ \(a,b) ->
-      rawSerialiseVerKeyKES a <> rawSerialiseVerKeyKES b
-
-slice :: Word -> Word -> ByteString -> ByteString
-slice offset size = BS.take (fromIntegral size)
-                  . BS.drop (fromIntegral offset)
-
-zeroSeed :: KESAlgorithm d => Proxy d -> Seed
-zeroSeed p = mkSeedFromBytes (BS.replicate seedSize (0 :: Word8))
-  where
-    seedSize :: Int
-    seedSize = fromIntegral (seedSizeKES p)
-
-mungeName :: String -> String
-mungeName basename
-  | (name, '^':nstr) <- span (/= '^') basename
-  , [(n, "")] <- reads nstr
-  = name ++ '^' : show (n+1 :: Word)
-
-  | otherwise
-  = basename ++ "_2^1"
 
 
 --
